requirements.txt
# Core RAG
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0

# Vector Store
faiss-cpu>=1.7.4

# SEC Filing Download
sec-edgar-downloader>=0.4.0

# HTTP
requests>=2.31.0

# Document Parsing
beautifulsoup4>=4.12.0
unstructured[html]>=0.13.0

# Vector Store (optional backend)
chromadb>=0.5.0
pypdf>=4.0.0
pdfplumber>=0.10.0

# Embeddings & NLP
sentence-transformers>=2.7.0
tiktoken>=0.7.0

# Data Handling
pandas>=2.0.0
numpy>=1.26.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Evaluation
rouge-score>=0.1.2
ragas>=0.1.0